Removed target="_blank"
RPOL version change caused the parser to no longer value the target.
Their developer believes the user should make the decision on how to
open links with CTRL,Shift,etc and target shouldnt be used.   I was the
only user of it so I guess it's not really up to the users, but up to
him to decide on how the user experience should work.  Jerks I tell ya!